Adjust modules structure, add necessary imports, improve naming.
